refactored server.js into multiple files and changed up some of the flow through the authentication process
